requirements.txt
Package         Version
--------------- -----------
narwhals        1.37.1
numpy           2.2.5
packaging       25.0
pandas          2.2.3
pillow          11.2.1
pip             22.0.2
plotly          6.0.1
PyQt5           5.15.11
PyQt5-Qt5       5.15.16
PyQt5_sip       12.17.0
PyQt6           6.9.0
PyQt6-Qt6       6.9.0
PyQt6_sip       13.10.0
python-dateutil 2.9.0.post0
pytz            2025.2
setuptools      59.6.0
six             1.17.0
tzdata          2025.2
